Created ProblemBlock and added it to a lesson + renamed lesson files
