+++ public/lessons/dynamic-programming/mcq-1.mdx
